fix: header cleanup, fix editor subsystem dependency
fix: additional macro condition for packaging
