+++ Source/wunthshin/Actors/Item/A_WSItem.h
@@ -69,7 +69,7 @@ public:
 	FName GetAssetName() const { return AssetName; }
 
 	virtual UClass* GetSubsystemType() const override;
-#ifdef WITH_EDITOR
+#if WITH_EDITOR & !UE_BUILD_SHIPPING_WITH_EDITOR 
 	virtual UClass* GetEditorSubsystemType() const override;
 #endif
