+++ Source/wunthshin/Actors/Item/Weapon/A_WSWeapon.h
@@ -31,7 +31,7 @@ public:
 	virtual UScriptStruct* GetTableType() const override;
 
 	virtual UClass* GetSubsystemType() const override;
-#ifdef WITH_EDITOR
+#if WITH_EDITOR & !UE_BUILD_SHIPPING_WITH_EDITOR 
 	virtual UClass* GetEditorSubsystemType() const override;
 #endif
