+++ Source/wunthshin/Components/Weapon/C_WSWeapon.cpp
@@ -4,13 +4,12 @@
 #include "C_WSWeapon.h"
 #include "InputMappingContext.h"
 #include "InputAction.h"
-#include "InputActionValue.h"
 #include "EnhancedInputSubsystems.h"
 #include "EnhancedInputComponent.h"
 #include "wunthshin/Components/PickUp/C_WSPickUp.h"
 #include "wunthshin/Actors/Item/Weapon/A_WSWeapon.h"
 #include "wunthshin/Interfaces/Taker/Taker.h"
-#include "wunthshin/Subsystem/WorldStatusSubsystem/WorldStatusSubsystem.h"
+#include "wunthshin/Subsystem/WorldSubsystem/WorldStatus/WorldStatusSubsystem.h"
 
 
 // Sets default values for this component's properties
