+++ Source/wunthshin/Data/Items/ItemMetadata/SG_WSItemMetadata.cpp
@@ -2,8 +2,8 @@
 
 
 #include "wunthshin/Data/Items/ItemMetadata/SG_WSItemMetadata.h"
-#include "wunthshin/Subsystem/EffectSubsystem/EffectSubsystem.h"
 #include "wunthshin/Data/Effects/EffectTableRow/EffectTableRow.h"
+#include "wunthshin/Subsystem/GameInstanceSubsystem/Effect/EffectSubsystem.h"
 
 const UO_WSBaseEffect* USG_WSItemMetadata::GetItemEffect(const UWorld* InWorld) const
 {
