+++ Source/wunthshin/Data/Items/ItemTableRow/ItemTableRow.cpp
@@ -5,8 +5,6 @@
 #include "Components/SphereComponent.h"
 
 #include "wunthshin/Data/Items/ItemMetadata/SG_WSItemMetadata.h"
-#include "wunthshin/Subsystem/ItemSubsystem/ItemSubsystem.h"
-#include "wunthshin/Subsystem/WeaponSubsystem/WeaponSubsystem.h"
 
 void FItemTableRow::OnDataTableChanged(const UDataTable* InDataTable, const FName InRowName)
 {
