+++ Source/wunthshin/Subsystem/GameInstanceSubsystem/NPC/NPCSubsystem.cpp
@@ -22,21 +22,3 @@ void UNPCSubsystem::Initialize(FSubsystemCollectionBase& Collection)
 	DataTableMapping.Emplace(FNPCTableRow::StaticStruct(), AssetDataTable);
 	DataTableMapping.Emplace(FNPCStats::StaticStruct(), StatDataTable);
 }
-
-void UNPCEditorSubsystem::Initialize(FSubsystemCollectionBase& Collection)
-{
-	Super::Initialize(Collection);
-	DataTableMapping.Emplace(FNPCTableRow::StaticStruct(), AssetDataTable);
-	DataTableMapping.Emplace(FNPCStats::StaticStruct(), StatDataTable);
-}
-
-UNPCEditorSubsystem::UNPCEditorSubsystem()
-{
-	static ConstructorHelpers::FObjectFinder<UDataTable> AssetTable(TEXT("/Script/Engine.DataTable'/Game/DataTable/DT_NPCTable.DT_NPCTable'"));
-	check(AssetTable.Object);
-	AssetDataTable = AssetTable.Object;
-	
-	static ConstructorHelpers::FObjectFinder<UDataTable> StatTable(TEXT("/Script/Engine.DataTable'/Game/DataTable/DT_NPCStatsTable.DT_NPCStatsTable'"));
-	check(StatTable.Object);
-	StatDataTable = StatTable.Object;
-}
