+++ Source/wunthshin/Subsystem/GameInstanceSubsystem/NPC/NPCSubsystem.h
@@ -28,21 +28,3 @@ public:
 	virtual void Initialize(FSubsystemCollectionBase& Collection) override;
 
 };
-
-UCLASS()
-class WUNTHSHIN_API UNPCEditorSubsystem : public UEditorSubsystem, public IDataTableQuery
-{
-	GENERATED_BODY()
-
-	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Table", meta=(AllowPrivateAccess = "true"))
-	UDataTable* AssetDataTable;
-
-	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Table", meta=(AllowPrivateAccess = "true"))
-	UDataTable* StatDataTable;
-	
-public:
-	UNPCEditorSubsystem();
-
-	virtual void Initialize(FSubsystemCollectionBase& Collection) override;
-
-};
