+++ Source/wunthshin/Subsystem/Utility.cpp
@@ -1,3 +1,1 @@
 #include "Utility.h"
-
-#include "wunthshin/Data/Items/ItemMetadata/SG_WSItemMetadata.h"
